requirements.txt
pip>=25.0.1
Flask
opencv-python
numpy
requests
torch>=2.1.0
torchvision==0.16.0
azure-storage-blob